+++ xclim/indices/_multivariate.py
@@ -3,6 +3,7 @@ from typing import Optional
 import numpy as np
 import xarray
 
+from . import fwi
 from xclim import run_length as rl
 from xclim import utils
 from xclim.utils import declare_units
@@ -26,6 +27,8 @@ __all__ = [
     "daily_temperature_range_variability",
     "days_over_precip_thresh",
     "extreme_temperature_range",
+    "fire_weather_indexes",
+    "drought_code",
     "fraction_over_precip_thresh",
     "heat_wave_frequency",
     "heat_wave_max_length",
@@ -327,6 +330,162 @@ def extreme_temperature_range(
     out = tx_max - tn_min
     out.attrs["units"] = f"{q.units:~}"
     return out
+
+
+@declare_units(
+    "",
+    check_output=False,
+    tas="[temperature]",
+    pr="[precipitation]",
+    ws="[speed]",
+    rh="[]",
+    snd="[length]",
+)
+def fire_weather_indexes(
+    tas: xarray.DataArray,
+    pr: xarray.DataArray,
+    ws: xarray.DataArray,
+    rh: xarray.DataArray,
+    lat: xarray.DataArray,
+    snd: xarray.DataArray = None,
+    ffmc0: xarray.DataArray = None,
+    dmc0: xarray.DataArray = None,
+    dc0: xarray.DataArray = None,
+    start_date: str = None,
+    **params,
+):
+    r"""Return the six daily fire weather indexes.
+
+    Computes the 6 fire weather indexes as defined by the Canadian Forest Service:
+    the Drought Code, the Duff-Moisture Code, the Fine Fuel Moisture Code,
+    the Initial Spread Index, the Build Up Index and the Fire Weather Index.
+
+    Parameters
+    ----------
+    tas : xarray.DataArray
+      Noon temperature.
+    pr : xarray.DataArray
+      Rain fall in open over previous 24 hours, at noon.
+    ws : xarray.DataArray
+      Noon wind speed.
+    rh : xarray.DataArray
+      Noon relative humidity.
+    lat : xarray.DataArray
+      Latitude coordinate
+    snd : xarray.DataArray
+      Noon snow depth.
+    ffmc0 : xarray.DataArray
+      Initial values of the fine fuel moisture code.
+    dmc0 : xarray.DataArray
+      Initial values of the Duff moisture code.
+    dc0 : xarray.DataArray
+      Initial values of the drought code.
+    start_date : str, datetime.datetime
+      Date at which to start the computation, dc0/dmc0/ffcm0 should be given at the day before.
+    params :
+        Any other keyword parameters as defined in `xclim.indices.fwi.fire_weather_ufunc`.
+
+    Returns
+    -------
+    DC, DMC, FFMC, ISI, BUI, FWI
+
+    Notes
+    -----
+    See https://cwfis.cfs.nrcan.gc.ca/background/dsm/fwi
+
+    References
+    ----------
+    Y. Wang, K.R. Anderson, and R.M. Suddaby, INFORMATION REPORT NOR-X-424, 2015.
+    """
+    tas = utils.convert_units_to(tas, "C")
+    pr = utils.convert_units_to(pr, "mm/day")
+    ws = utils.convert_units_to(ws, "km/h")
+    rh = utils.convert_units_to(rh, "pct")
+    if snd is not None:
+        snd = utils.convert_units_to(snd, "m")
+
+    if dc0 is None:
+        dc0 = xarray.full_like(tas.isel(time=0)) * np.nan
+    if dmc0 is None:
+        dc0 = xarray.full_like(tas.isel(time=0)) * np.nan
+    if ffmc0 is None:
+        dc0 = xarray.full_like(tas.isel(time=0)) * np.nan
+
+    params["start_date"] = start_date
+
+    out = fwi.fire_weather_ufunc(
+        tas=tas,
+        pr=pr,
+        rh=rh,
+        ws=ws,
+        lat=lat,
+        dc0=dc0,
+        dmc0=dmc0,
+        ffmc0=ffmc0,
+        snd=snd,
+        indices=["DC", "DMC", "FFMC", "ISI", "BUI", "FWI"],
+        **params,
+    )
+    return out["DC"], out["DMC"], out["FFMC"], out["ISI"], out["BUI"], out["FWI"]
+
+
+@declare_units("", tas="[temperature]", pr="[precipitation]", snd="[length]")
+def drought_code(
+    tas: xarray.DataArray,
+    pr: xarray.DataArray,
+    lat: xarray.DataArray,
+    snd: xarray.DataArray = None,
+    dc0: xarray.DataArray = None,
+    start_date: str = None,
+    **params,
+):
+    r"""Return the daily drought code
+
+    The drought code is part of the Canadian Forest Fire Weather Index System. It is a numeric rating of the average moisture content of organic layers.
+
+    Parameters
+    ----------
+    tas : xarray.DataArray
+      Noon temperature.
+    pr : xarray.DataArray
+      Rain fall in open over previous 24 hours, at noon.
+    lat : xarray.DataArray
+      Latitude coordinate
+    snd : xarray.DataArray
+      Noon snow depth.
+    dc0 : xarray.DataArray
+      Initial values of the drought code.
+    start_date : str, datetime.datetime
+      Date at which to start the computation, dc0/dmc0/ffcm0 should be given at the day before.
+    params :
+      Any other keyword parameters as defined in `xclim.indices.fwi.fire_weather_ufunc`.
+
+    Returns
+    -------
+    Drought code [-]
+
+    Notes
+    -----
+    See https://cwfis.cfs.nrcan.gc.ca/background/dsm/fwi
+
+    References
+    ----------
+    Y. Wang, K.R. Anderson, and R.M. Suddaby, INFORMATION REPORT NOR-X-424, 2015.
+    """
+    tas = utils.convert_units_to(tas, "C")
+    pr = utils.convert_units_to(pr, "mm/day")
+    if snd is not None:
+        snd = utils.convert_units_to(snd, "m")
+
+    if dc0 is None:
+        dc0 = xarray.full_like(tas.isel(time=0), np.nan)
+
+    params["start_date"] = start_date
+
+    out = fwi.fire_weather_ufunc(
+        tas=tas, pr=pr, lat=lat, dc0=dc0, snd=snd, indexes=["DC"], **params
+    )
+    return out["DC"]
 
 
 @declare_units(