+++ xclim/subset.py
@@ -1,4 +1,5 @@
 import copy
+import logging
 import warnings
 from functools import wraps
 from pathlib import Path
@@ -164,6 +165,55 @@ def check_lons(func):
     return func_checker
 
 
+def check_latlon_dimnames(func):
+    @wraps(func)
+    def func_checker(*args, **kwargs):
+        """
+        A decorator examining the names of the latitude and longitude dimensions and renames them temporarily.
+         Checks here ensure that the names supplied via the xarray object dims are changed to be synonymous with subset
+         algorithm dimensions, conversions are saved and are then undone to the processed file.
+        """
+
+        if range(len(args)) == 0:
+            return func(*args, **kwargs)
+
+        formatted_args = list()
+        conv = dict()
+        for argument in args:
+            if isinstance(argument, (xarray.DataArray, xarray.Dataset)):
+                dims = argument.dims
+            else:
+                logging.info(f"No file or no dimensions found in arg `{argument}`.")
+                formatted_args.append(argument)
+                continue
+
+            if not {"lon", "lat"}.issubset(dims):
+                if {"long"}.issubset(dims):
+                    conv["long"] = "lon"
+                elif {"latitude", "longitude"}.issubset(dims):
+                    conv["latitude"] = "lat"
+                    conv["longitude"] = "lon"
+                elif {"lats", "lons"}.issubset(dims):
+                    conv["lats"] = "lat"
+                    conv["lons"] = "lon"
+                if not conv and not {"rlon", "rlat"}.issubset(dims):
+                    warnings.warn(
+                        f"lat and lon-like dimensions are not found among arg `{argument}` dimensions: {list(dims)}."
+                    )
+                argument = argument.rename(conv)
+
+            formatted_args.append(argument)
+
+        final = func(*formatted_args, **kwargs)
+
+        for k, v in conv.items():
+            final = final.rename({v: k})
+
+        return final
+
+    return func_checker
+
+
 def wrap_lons_and_split_at_greenwich(func):
     @wraps(func)
     def func_checker(*args, **kwargs):
@@ -193,37 +243,38 @@ def wrap_lons_and_split_at_greenwich(func):
                     stacklevel=4,
                 )
             split_flag = False
-            if (poly.geometry.total_bounds[0] < 0) and (
-                poly.geometry.total_bounds[2] > 0
-            ):
-                split_flag = True
-                warnings.warn(
-                    "Geometry crosses the Greenwich Meridian. Proceeding to split polygon at Greenwich."
-                    " This feature is experimental. Output might not be accurate.",
-                    UserWarning,
-                    stacklevel=4,
-                )
+            for (index, feature) in poly.iterrows():
+                if (feature.geometry.bounds[0] < 0) and (
+                    feature.geometry.bounds[2] > 0
+                ):
+                    split_flag = True
+                    warnings.warn(
+                        "Geometry crosses the Greenwich Meridian. Proceeding to split polygon at Greenwich."
+                        " This feature is experimental. Output might not be accurate.",
+                        UserWarning,
+                        stacklevel=4,
+                    )
 
-                # Create a meridian line at Greenwich, split polygons at this line and erase a buffer line
-                union = Polygon(cascaded_union(poly.geometry))
-                meridian = LineString([Point(0, 90), Point(0, -90)])
-                buffered = meridian.buffer(0.000000001)
-                split_polygons = split(union, meridian)
-                # TODO: This doesn't seem to be thread safe in Travis CI on macOS. Merits testing with a local machine.
-                buffered_split_polygons = [
-                    feat for feat in split_polygons.difference(buffered)
-                ]
+                    # Create a meridian line at Greenwich, split polygons at this line and erase a buffer line
+                    union = Polygon(cascaded_union(feature.geometry))
+                    meridian = LineString([Point(0, 90), Point(0, -90)])
+                    buffered = meridian.buffer(0.000000001)
+                    split_polygons = split(union, meridian)
+                    # TODO: This doesn't seem to be thread safe in Travis CI on macOS. Merits testing with a local machine.
+                    buffered_split_polygons = [
+                        feat for feat in split_polygons.difference(buffered)
+                    ]
 
-                # Load split features into a new GeoDataFrame with WGS84 CRS
-                split_gdf = gpd.GeoDataFrame(
-                    list(range(len(buffered_split_polygons))),
-                    geometry=buffered_split_polygons,
-                    crs={"init": "epsg:4326"},
-                )
-                # split_gdf.crs = CRS.from_epsg(4326)
-                split_gdf.columns = ["index", "geometry"]
+                    # Cannot assign iterable with `at` (pydata/pandas#26333) so a small hack:
+                    # Load split features into a new GeoDataFrame with WGS84 CRS
+                    split_gdf = gpd.GeoDataFrame(
+                        geometry=[cascaded_union(buffered_split_polygons)],
+                        crs={"epsg:4326"},
+                    )
+                    poly.at[[index], "geometry"] = split_gdf.geometry.values
+                    # split_gdf.columns = ["index", "geometry"]
 
-                poly = split_gdf
+                    # feature = split_gdf
 
             # Reproject features in WGS84 CSR to use 0 to 360 as longitudinal values
             poly = poly.to_crs(
@@ -238,6 +289,7 @@ def wrap_lons_and_split_at_greenwich(func):
                 )
                 poly = gpd.GeoDataFrame(poly.buffer(0.000000001), columns=["geometry"])
                 poly.crs = crs1
+
             kwargs["poly"] = poly
 
         return func(*args, **kwargs)
@@ -253,7 +305,10 @@ def create_mask(
     poly: gpd.GeoDataFrame = None,
     wrap_lons: bool = False,
 ):
-    """
+    """Creates a mask with values corresponding to the features in a GeoDataFrame.
+
+    The returned mask's points have the value of the first geometry of `poly` they fall in.
+
     Parameters
     ----------
     x_dim : xarray.DataArray
@@ -268,7 +323,33 @@ def create_mask(
     Returns
     -------
     xarray.DataArray
+
+    Examples
+    --------
+    >>> from xclim import subset
+    >>> import xarray as xr
+    >>> import geopandas as gpd
+    >>> ds = xr.open_dataset('example.nc')
+    >>> polys = gpd.read_file('regions.json')
+    Get a mask from all polygons in 'regions.json'
+    >>> mask = subset.create_mask(x_dim=ds.lon, y_dim=ds.lat, poly=polys)
+    >>> ds = ds.assign_coords(regions=mask)
+    Operations can be applied to each regions with  `groupby`. Ex:
+    >>> ds = ds.groupby('regions').mean()
+    Extra step to retrieve the names of those polygons stored in the "id" column
+    >>> region_names = xr.DataArray(polys.id, dims=('regions',)))
+    >>> ds = ds.assign_coords(regions_names=region_names)
     """
+    # Check for intersections
+    for i, (inda, pola) in enumerate(poly.iterrows()):
+        for (indb, polb) in poly.loc[i + 1 :].iterrows():
+            if pola.geometry.intersects(polb.geometry):
+                warnings.warn(
+                    f"List of shapes contains overlap between {inda} and {indb}. Only {inda} will be used.",
+                    UserWarning,
+                    stacklevel=4,
+                )
+
     if len(x_dim.shape) == 1 & len(y_dim.shape) == 1:
         # create a 2d grid of lon, lat values
         lon1, lat1 = np.meshgrid(
@@ -313,6 +394,7 @@ def create_mask(
     return mask_2d
 
 
+@check_latlon_dimnames
 def subset_shape(
     ds: Union[xarray.DataArray, xarray.Dataset],
     shape: Union[str, Path],
@@ -474,6 +556,7 @@ def subset_shape(
     return ds_copy
 
 
+@check_latlon_dimnames
 @check_lons
 @check_date_signature
 def subset_bbox(
@@ -653,6 +736,7 @@ def _check_desc_coords(coord, bounds, dim):
     return bounds
 
 
+@check_latlon_dimnames
 @check_lons
 @check_date_signature
 def subset_gridpoint(