+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.14.x
+------
+* Fix in `subset.wrap_lons_and_split_at_greenwich` to preserve multi-region dataframes.
+* Improve the memory use of `indices.growing_season_length`.
+* Better handling of data with atypically named `lat` and `lon` dimensions.
+* Added 6 Fire Weather Indexes.
+
 0.13.x (2020-01-10)
 -------------------
 * Documentation improvements: list of indicators, RTD theme, notebook example.